requirements.txt
Flask
flask_cors
psycopg2-binary
gunicorn